+++ attributes/default.rb
@@ -31,7 +31,7 @@ default['cacti']['db_file'] = value_for_platform(
   %w(pld) => {
     'default' => '/etc/webapps/cacti/config.php'
   },
-  %w(ubuntu) => {
+  %w(debian ubuntu) => {
     'default' => '/etc/cacti/debian.php'
   }
 )
@@ -39,10 +39,15 @@ default['cacti']['group'] = value_for_platform(
   %w(pld) => {
     'default' => 'http'
   },
-  %w(centos fedora redhat ubuntu) => {
+  %w(centos fedora redhat) => {
     'default' => 'apache'
+  },
+  %w(debian ubuntu) => {
+    'default' => 'www-data'
   }
 )
+
+default['cacti']['mysql_provider'] = 'mysql'
 
 default['cacti']['packages'] = value_for_platform(
   %w(centos fedora redhat) => {
@@ -51,7 +56,7 @@ default['cacti']['packages'] = value_for_platform(
   %w(pld) => {
     'default' => %w(cacti cacti-setup)
   },
-  %w(ubuntu) => {
+  %w(debian ubuntu) => {
     %w(12.04 12.10 13.04) => %w(cacti libsnmp-base libsnmp15 snmp snmpd libnet-ldap-perl libnet-snmp-perl php-net-ldap php5-mysql php-apc php5-snmp),
     '13.10' => %w(cacti libsnmp-base libsnmp30 snmp snmpd libnet-ldap-perl libnet-snmp-perl php-net-ldap php5-mysql php-apc php5-snmp),
     'default' => %w(cacti libsnmp-base libsnmp30 snmp snmpd libnet-ldap-perl libnet-snmp-perl php-net-ldap php5-mysql php-apc php5-snmp)
@@ -64,7 +69,7 @@ default['cacti']['poller_file'] = value_for_platform(
   %w(pld) => {
     'default' => '/usr/sbin/cacti-poller'
   },
-  %w(ubuntu) => {
+  %w(debian ubuntu) => {
     'default' => '/usr/share/cacti/site/poller.php'
   }
 )
@@ -72,7 +77,7 @@ default['cacti']['poller_cmd'] = value_for_platform(
   %w(pld) => {
     'default' => "umask 022; exec #{node['cacti']['poller_file']} >> /var/log/cacti/poller.log 2>&1"
   },
-  %w(centos fedora redhat ubuntu) => {
+  %w(centos debian fedora redhat ubuntu) => {
     'default' => "/usr/bin/php #{node['cacti']['poller_file']} > /dev/null 2>&1"
   }
 )
@@ -90,6 +95,9 @@ default['cacti']['version'] = value_for_platform(
     '12.04' => '0.8.7i',
     %w(12.10 13.04) => '0.8.8a',
     '13.10' => '0.8.8b',
+    'default' => '0.8.8b'
+  },
+  %w(debian) => {
     'default' => '0.8.8a'
   }
 )
@@ -108,8 +116,11 @@ default['cacti']['apache2']['ssl']['certificate_file'] = value_for_platform(
   %w(pld) => {
     'default' => '/etc/httpd/ssl/server.crt'
   },
-  %w(centos fedora redhat ubuntu) => {
+  %w(centos fedora redhat) => {
     'default' => '/etc/pki/tls/certs/localhost.crt'
+  },
+  %w(debian ubuntu) => {
+    'default' => '/etc/ssl/certs/ssl-cert-snakeoil.pem'
   }
 )
 
@@ -120,8 +131,11 @@ default['cacti']['apache2']['ssl']['key_file'] = value_for_platform(
   %w(pld) => {
     'default' => '/etc/httpd/ssl/server.key'
   },
-  %w(centos fedora redhat ubuntu) => {
+  %w(centos fedora redhat) => {
     'default' => '/etc/pki/tls/private/localhost.key'
+  },
+  %w(debian ubuntu) => {
+    'default' => '/etc/ssl/private/ssl-cert-snakeoil.key'
   }
 )
 
@@ -147,7 +161,7 @@ default['cacti']['rrdtool']['version'] = value_for_platform(
     '2.0' => '1.2',
     'default' => '1.4'
   },
-  %w(ubuntu) => {
+  %w(debian ubuntu) => {
     'default' => '1.4'
   }
 )
@@ -163,7 +177,7 @@ default['cacti']['spine']['packages'] = value_for_platform(
   %w(pld) => {
     'default' => %w(net-snmp-devel openssl-devel)
   },
-  %w(ubuntu) => {
+  %w(debian ubuntu) => {
     'default' => %w(libsnmp-dev libssl-dev)
   }
 )